requirements.txt
asn1crypto==0.24.0
cffi==1.12.1
cryptography==2.5
cycler==0.10.0
Django==2.1.5
djangorestframework==3.9.4
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.0
opencv-python==4.1.2.30
pycparser==2.19
PyMySQL==0.9.3
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
Pillow==7.0.0
pyproj==2.6.1